requirements.txt
telethon>=1.36
aiohttp>=3.9
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
pyyaml>=6.0
loguru>=0.7
